requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai>=1.12.0
reportlab==4.1.0
python-dotenv==1.0.0

# v2.0 dependencies
easyocr==1.7.1
Pillow==10.2.0
numpy==1.26.3

# Optional (for name detection)
spacy==3.7.2
# After install: python -m spacy download en_core_web_sm

